Added test for the panic from the past
