+++ tests/test-helpers/src/lib.rs
@@ -1,2 +1,11 @@
 pub mod builder;
 pub mod fuel_core_driver;
+
+pub async fn send_graph_ql_query(url: &str, query: &str) -> String {
+    let client = reqwest::Client::new();
+    let mut map = std::collections::HashMap::new();
+    map.insert("query", query);
+    let response = client.post(url).json(&map).send().await.unwrap();
+
+    response.text().await.unwrap()
+}
