+++ tests/tests/dos.rs
@@ -4,15 +4,7 @@ use fuel_core::service::{
     Config,
     FuelService,
 };
-
-async fn send_query(url: &str, query: &str) -> String {
-    let client = reqwest::Client::new();
-    let mut map = std::collections::HashMap::new();
-    map.insert("query", query);
-    let response = client.post(url).json(&map).send().await.unwrap();
-
-    response.text().await.unwrap()
-}
+use test_helpers::send_graph_ql_query;
 
 #[tokio::test]
 async fn complex_queries__recursion() {
@@ -81,7 +73,7 @@ async fn complex_queries__recursion() {
     let node = FuelService::new_node(config).await.unwrap();
     let url = format!("http://{}/v1/graphql", node.bound_address);
 
-    let result = send_query(&url, query).await;
+    let result = send_graph_ql_query(&url, query).await;
     assert!(result.contains("The recursion depth of the query cannot be greater than"));
 }
 
@@ -102,7 +94,7 @@ async fn complex_queries__10_blocks__works() {
     let node = FuelService::new_node(Config::local_node()).await.unwrap();
     let url = format!("http://{}/v1/graphql", node.bound_address);
 
-    let result = send_query(&url, query).await;
+    let result = send_graph_ql_query(&url, query).await;
     assert!(result.contains("transactions"));
 }
 
@@ -123,6 +115,6 @@ async fn complex_queries__50_block__query_to_complex() {
     let node = FuelService::new_node(Config::local_node()).await.unwrap();
     let url = format!("http://{}/v1/graphql", node.bound_address);
 
-    let result = send_query(&url, query).await;
+    let result = send_graph_ql_query(&url, query).await;
     assert!(result.contains("Query is too complex."));
 }
